requirements.txt
Babel==2.3.4
ConfigArgParse==0.10.0
Django==1.9.4
Flask==0.11.1
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==3.3.0
PyYAML==3.11
Werkzeug==0.11.10
amqp==1.4.9
anyjson==0.3.3
argh==0.26.2
argparse==1.2.1
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
billiard==3.3.0.23
blinker==1.4
celery==3.1.23
certifi==2016.2.28
cffi==1.7.0
click==6.6
construct==2.5.2
cryptography==1.4
django-celery==3.1.17
django-pagination==1.0.7
enum34==1.1.6
flower==0.9.1
futures==3.0.5
gevent==1.1.1
greenlet==0.4.10
h2==2.4.0
hpack==2.2.0
html2text==2016.5.29
hyperframe==4.0.1
idna==2.1
ipaddress==1.0.16
itsdangerous==0.24
kombu==3.0.35
lxml==3.6.0
mitmproxy==0.18
netlib==0.17
passlib==1.6.5
pathtools==0.1.2
psutil==4.3.0
pyOpenSSL==16.0.0
pyasn1==0.1.9
pycparser==2.14
pyparsing==2.1.5
pyperclip==1.5.27
pytz==2016.3
redis==2.10.5
requests==2.10.0
singledispatch==3.4.0.3
six==1.10.0
tornado==4.3
typing==3.5.2.2
urwid==1.3.1
watchdog==0.8.3
wsgiref==0.1.2

